Condense cmprovider to a single package
