--- cmprovider/api/src/java/org/sakaiproject/cmprovider/api/data/annotations/After.java
@@ -1,17 +1,0 @@
-package org.sakaiproject.cmprovider.api.data.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Annotation to enforce order of date or formatted date string fields.
- *
- * @author Christopher Schauer
- */
-@Retention(value=RetentionPolicy.RUNTIME)
-@Target(value=ElementType.FIELD)
-public @interface After {
-  String field();
-}
