--- cmprovider/api/src/java/org/sakaiproject/cmprovider/api/data/annotations/NotNull.java
@@ -1,16 +1,0 @@
-package org.sakaiproject.cmprovider.api.data.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Annotation to be used on a field that should not be null.
- *
- * @author Christopher Schauer
- */
-@Retention(value=RetentionPolicy.RUNTIME)
-@Target(value=ElementType.FIELD)
-public @interface NotNull {
-}
