--- cmprovider/pack/pom.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0">
-  <modelVersion>4.0.0</modelVersion>
-  <name>Course Management Provider Component Package</name>
-  <parent>
-    <groupId>org.sakaiproject.cmprovider</groupId>
-    <artifactId>cmprovider-base</artifactId>
-    <version>11-SNAPSHOT</version>
-    <relativePath>../pom.xml</relativePath>
-  </parent>
-  <groupId>org.sakaiproject.cmprovider</groupId>
-  <artifactId>cmprovider-pack</artifactId>
-  <organization>
-    <name>Texas State University</name>
-    <url>http://txstate.edu/</url>
-  </organization>
-  <inceptionYear>2014</inceptionYear>
-  <packaging>sakai-component</packaging>
-  <properties>
-    <deploy.target>components</deploy.target>
-  </properties>
-  <dependencies>
-    <dependency>
-      <groupId>org.sakaiproject.cmprovider</groupId>
-      <artifactId>cmprovider-impl</artifactId>
-      <version>11-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.sakaiproject.cmprovider</groupId>
-      <artifactId>cmprovider-api</artifactId>
-      <version>11-SNAPSHOT</version>
-      <scope>provided</scope>
-    </dependency>
-  </dependencies>
-  <build>
-    <resources>
-      <resource>
-        <directory>${basedir}/src/bundle</directory>
-        <includes>
-          <include>*.properties</include>
-        </includes>
-      </resource>
-    </resources>
-  </build>
-</project>
