+++ cmprovider/src/webapp/WEB-INF/components.xml
